test: split tournament e2e tests
