--- src/http/controllers/tournaments/tournaments.e2e.spec.ts
@@ -1,114 +1,0 @@
-import { Tournament, Match } from '@prisma/client';
-import request from 'supertest';
-import { beforeAll, describe, expect, it } from 'vitest';
-
-import { IMatchesRepository } from '@/repositories/matches/IMatchesRepository';
-import { PrismaMatchesRepository } from '@/repositories/matches/PrismaMatchesRepository';
-import { ITeamsRepository } from '@/repositories/teams/ITeamsRepository';
-import { PrismaTeamsRepository } from '@/repositories/teams/PrismaTeamsRepository';
-import { ITournamentsRepository } from '@/repositories/tournaments/ITournamentsRepository';
-import { PrismaTournamentsRepository } from '@/repositories/tournaments/PrismaTournamentsRepository';
-import { createTestApp } from '@/test/helper-e2e';
-import { createMatchWithTeams } from '@/test/mocks/match';
-import { createTournament } from '@/test/mocks/tournament';
-
-type TournamentResponse = {
-  tournaments: Array<Tournament>;
-};
-
-type TournamentMatchesResponse = {
-  matches: Array<Match>;
-};
-
-describe('Tournaments E2E', async () => {
-  const app = await createTestApp();
-  let token: string;
-  let tournamentId: number;
-
-  let tournamentsRepository: ITournamentsRepository;
-  let matchesRepository: IMatchesRepository;
-  let teamsRepository: ITeamsRepository;
-
-  beforeAll(async () => {
-    token = app.jwt.sign({ sub: 'test-user' });
-    tournamentsRepository = new PrismaTournamentsRepository();
-    matchesRepository = new PrismaMatchesRepository();
-    teamsRepository = new PrismaTeamsRepository();
-
-    // Create a tournament for testing
-    const tournament = await createTournament(tournamentsRepository, {});
-    tournamentId = tournament.id;
-
-    // Create matches for the tournament
-    for (let i = 1; i <= 3; i++) {
-      await createMatchWithTeams({ matchesRepository, teamsRepository }, { tournamentId });
-    }
-  });
-
-
-  describe('GET /tournaments', () => {
-    it('should be able to list all tournaments', async () => {
-      const response = await request(app.server)
-        .get('/tournaments')
-        .set('Authorization', `Bearer ${token}`);
-
-      expect(response.status).toBe(200);
-      const body = response.body as TournamentResponse;
-      expect(body.tournaments).toBeInstanceOf(Array);
-      expect(body.tournaments.length).toBeGreaterThanOrEqual(1);
-    });
-
-    it('should require authentication', async () => {
-      const response = await request(app.server).get('/tournaments');
-
-      expect(response.status).toBe(401);
-    });
-
-    it('should return 422 when query params are invalid', async () => {
-      const response = await request(app.server)
-        .get('/tournaments')
-        .query({ limit: 'invalid' })
-        .set('Authorization', `Bearer ${token}`);
-
-      expect(response.status).toBe(422);
-    });
-  });
-
-  describe('GET /tournaments/:tournamentId/matches', () => {
-    it('should be able to get matches for a tournament', async () => {
-      const response = await request(app.server)
-        .get(`/tournaments/${tournamentId}/matches`)
-        .set('Authorization', `Bearer ${token}`);
-
-      expect(response.status).toBe(200);
-      const body = response.body as TournamentMatchesResponse;
-      expect(body.matches).toBeInstanceOf(Array);
-      expect(body.matches.length).toBe(3);
-    });
-
-    it('should require authentication', async () => {
-      const response = await request(app.server).get(
-        `/tournaments/${tournamentId}/matches`
-      );
-
-      expect(response.status).toBe(401);
-    });
-
-    it('should return 404 for non-existent tournament', async () => {
-      const response = await request(app.server)
-        .get(`/tournaments/999999/matches`)
-        .set('Authorization', `Bearer ${token}`);
-
-      expect(response.status).toBe(404);
-    });
-
-    it('should return 422 when query params are invalid', async () => {
-      const response = await request(app.server)
-        .get(`/tournaments/${tournamentId}/matches`)
-        .query({ limit: 0 })
-        .set('Authorization', `Bearer ${token}`);
-
-      expect(response.status).toBe(422);
-    });
-  });
-});
